feat: tag CloudFormation stacks deployed by CLI and only allow tagged resources to list and delete #35
This change allows to tag CloudFormation stacks that are deployed via CLI with "aws-dotnet-deploy" tag key. The same tag key is used to filter the CloudFormation stacks while using the list command.

Delete command also updated so that only the stacks that are deployed via the CLI can be deleted with delete command.