feat: add dependency injection using .NET Core DI framework
This change introduce dependency injection using .NET Core DI framework. Most the code remain as it is except the dependency creation phase. CustomServiceCollectionExtension is responsible for creating all the dependencies and add them to serviceCollection.

Responsibility of Program.cs is broken down to App.cs and CommandFactory without any change in functionality.