feat: Add support for custom Dockerfile paths
You can now use the DockerfilePath advanced setting to specify a path to
a Dockerfile as an alternative to the default location and name for
Docker-based recipes.

BREAKING CHANGE: Previously DockerExecutionDirectory was treating paths
as relative to the directory where the deployment tool is running.
If you were specifying a relative Docker execution directory, adjust
it to be relative to the project path.