fix: ask for "Create New" even if options to select is empty
## Motivation
`AskUserToChooseOrCreateNew` allows users to select an existing option or create new one. When there are no options to select, **Create New** is not shown and CLI returns to previous screen with default selection **Create New** which is correct in theory because that's the only action user can take. However, this is not a good UX.

## Changes
The check to display selection is moved to (Options + ** Empty** + **Create New**) combination, i.e. if any of the options are present, user will be asked to pick one. If none of the exists, the there is modeling error and should be handled at recipe level.

## Result
In cases like selecting ECS application role and there doesn't exist any compatible role in AWS account, customer is still asked to picked Create New.