fix: catch AmazonServiceException on making calls to AWS services and wrap to DeployToolException
## Motivation
When deploy tool makes call to AWS services and a service call fails, this is shown as bug to customer which in most cases are configuration issues such as not having correct permissions for the operation call made. This gives a false impression to customer that there a bug in the deploy tool.

## Changes
Similar to other places, this change wraps the AWS service calls to a try-catch block and if the exception is `AmazonServiceException` which is base exception for all operations, it is wrapped in a  `ResourceQueryException` i.e. a `DeployToolException` to make it an expected exception. Expected exceptions don't emit stacktrace unless `--diagnostics` is provided.

## Result
Case such as not having permission to make a operation call now, shows a ErrorCode and message instead showing stacktrace. On passing the `--diagnostics` flag, stacktrace is shown for debug purpose.